# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Book Digitization Platform API",
    "description": "AI-powered API for document digitization, multi-tenant management, and RAG-based semantic search.",
    "version": "1.0.0",
    "contact": {
      "name": "Dynasty Empire Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Meta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "total": { "type": "integer", "example": 100 },
          "limit": { "type": "integer", "example": 20 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "insufficient_permissions" },
          "message": { "type": "string", "example": "Scholar role cannot delete books" },
          "details": { "type": "object" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "scholar", "professional", "guest"] },
          "tenantId": { "type": "string", "format": "uuid" }
        }
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "example": "Summa Theologica" },
          "author": { "type": "string", "example": "Thomas Aquinas" },
          "isbn": { "type": "string", "example": "978-3-16-148410-0" },
          "pageCount": { "type": "integer", "example": 450 }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "pageNumber": { "type": "integer", "example": 1 },
          "textContent": { "type": "string" },
          "imageUrl": { "type": "string", "format": "uri" }
        }
      },
      "RagResult": {
        "type": "object",
        "properties": {
          "text": { "type": "string" },
          "score": { "type": "number", "format": "float" },
          "bookId": { "type": "string", "format": "uuid" },
          "pageNumber": { "type": "integer" }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new foundation user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "tenantId", "role"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 12 },
                  "tenantId": { "type": "string", "format": "uuid" },
                  "role": { "type": "string", "enum": ["admin", "scholar", "professional", "guest"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": { "type": "string", "format": "uuid" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/sign-in": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate user and return session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/books": {
      "get": {
        "tags": ["Book Management"],
        "summary": "List all digitized books for tenant",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search by title or author" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": {
          "200": {
            "description": "List of books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Book" } },
                    "meta": { "$ref": "#/components/schemas/Meta" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/{id}": {
      "get": {
        "tags": ["Book Management"],
        "summary": "Get metadata for a specific book",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Book details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Book" }
                  }
                }
              }
            }
          },
          "404": { "description": "Book not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/books/{id}/pages": {
      "get": {
        "tags": ["Book Management"],
        "summary": "Get all page texts/images for a book",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "List of pages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Page" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/rag": {
      "post": {
        "tags": ["AI Search & Retrieval"],
        "summary": "Perform semantic RAG search",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string", "example": "What does Aquinas say about grace?" },
                  "top_k": { "type": "integer", "default": 5 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/RagResult" } }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}